hypothesis testing improvements
